+++ CHANGES.rst
@@ -19,6 +19,7 @@ in progress
 - Wording: Use ``MomentName`` for generator
 - Wording: Rename ``--wordlength`` to ``--size``
 - Swap display order of generators
+- Generate multiple tokens at once
 
 
 2019-06-28 0.0.0